rename subnetmgr to manager
