+++ chain/consensus/hierarchical/subnet/manager/api.go
@@ -1,4 +1,4 @@
-package subnetmgr
+package manager
 
 import (
 	"context"
