+++ chain/consensus/hierarchical/subnet/manager/crossmsg.go
@@ -1,4 +1,4 @@
-package subnetmgr
+package manager
 
 import (
 	"context"
