+++ chain/consensus/hierarchical/subnet/manager/crossmsg_test.go
@@ -1,4 +1,4 @@
-package subnetmgr
+package manager
 
 import (
 	"testing"
