+++ chain/consensus/hierarchical/subnet/manager/events.go
@@ -1,4 +1,4 @@
-package subnetmgr
+package manager
 
 import (
 	"context"
