+++ chain/consensus/hierarchical/subnet/manager/exchange.go
@@ -1,4 +1,4 @@
-package subnetmgr
+package manager
 
 import (
 	"context"
